requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-pptx==0.6.23
python-multipart==0.0.6
Pillow==10.1.0
requests==2.31.0
google-generativeai==0.3.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
lxml==5.1.0
Pillow==10.1.0
